requirements.txt
requests
pyquery
python-dotenv
